eCommercePlatform/requirements.txt
aiohappyeyeballs==2.3.6
aiohttp==3.10.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.1
cron-descriptor==1.4.3
cryptography==43.0.0
dj-stripe==2.8.4
Django==5.0.7
django-admin-interface==0.28.8
django-allauth==64.1.0
django-bootstrap5==24.2
django-celery-beat==2.6.0
django-colorfield==0.11.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-filter==24.2
django-jsonfield==1.4.1
django-redis==5.4.0
django-stripe-payments==2.0.0
django-timezone-field==7.0
djangorestframework==3.15.2
environs==11.0.0
frozenlist==1.4.1
gunicorn>=20.0.0
idna==3.7
kombu==5.4.0
marshmallow==3.22.0
multidict==6.0.5
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
pycparser==2.22
PyJWT==2.9.0
pymemcache==4.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
redis==5.0.8
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
stripe==4.2.0
text-unidecode==1.3
twilio==9.2.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
